refactor disk logic for exec_contect

+++ lib/daemons/sequreispd.rb
@@ -48,29 +48,41 @@ def check_squid
   end
 
   #2: swap.state check
-  max_swap_size=500*1024*1024
-  cache_disk = Disk.cache.first
-  swap_file= cache_disk.nil? ? "/var/spool/squid/swap.state" : "/mnt/sequreisp#{cache_disk.name}/squid/swap.state"
-  swap_size = File.open(swap_file) do |sf| sf.size end
-  if swap_size > max_swap_size
-    Rails.logger.warn "sequreispd: #{Time.now.to_s} swap.state bigger than max: #{max_swap_size}, current: #{swap_size}, killing"
-    max_sleep = 20
-    while pid.present? and max_sleep > 0
-      system "kill -9 #{pid}"
-      sleep 1
-      pid = squid_pids
-      max_sleep -= 1
+  begin
+    max_swap_size=500*1024*1024
+    # swap_files = `find /var/spool/squid | grep 'swap.state$'`.split
+    swap_files = `grep "^cache_dir*" /etc/squid/squid.conf`.split("\n")
+    swap_files = `grep "^cache_dir*" /etc/squid/sequreisp.squid.conf`.split("\n") if swap_files.empty?
+    # swap_file= "/var/spool/squid/swap.state"
+    unless swap_files.empty?
+      swap_files.each do |line|
+        swap_file = "#{line.split[2]}/swap.state"
+        swap_size = File.open(swap_file) do |sf| sf.size end
+        if swap_size > max_swap_size
+          Rails.logger.warn "sequreispd: #{Time.now.to_s} swap.state bigger than max: #{max_swap_size}, current: #{swap_size}, killing"
+          max_sleep = 20
+          while pid.present? and max_sleep > 0
+            system "kill -9 #{pid}"
+            sleep 1
+            pid = squid_pids
+            max_sleep -= 1
+          end
+          if max_sleep == 0
+            Rails.logger.error "sequreispd: #{Time.now.to_s} could not kill squid after 20 tries, aborting"
+          else
+            Rails.logger.error "sequreispd: #{Time.now.to_s} starting squid"
+            old_swap_file="#{swap_file}.old"
+            FileUtils.mv swap_file, old_swap_file
+            system "/usr/sbin/service squid start"
+            FileUtils.rm old_swap_file
+          end
+        end
+      end
     end
-    if max_sleep == 0
-      Rails.logger.error "sequreispd: #{Time.now.to_s} could not kill squid after 20 tries, aborting"
-    else
-      Rails.logger.error "sequreispd: #{Time.now.to_s} starting squid"
-      old_swap_file="#{swap_file}.old"
-      FileUtils.mv swap_file, old_swap_file
-      system "/usr/sbin/service squid start"
-      FileUtils.rm old_swap_file
-    end
+  rescue => e
+    Rails.logger.error "ERROR check_squid swap.state check: #{e.inspect}"
   end
+
 
   #3: load average check
   max_load_average = Configuration.transparent_proxy_max_load_average
